Fix IDE0004

+++ src/libraries/System.Speech/src/Internal/SrgsCompiler/CfgSemanticTag.cs
@@ -75,7 +75,7 @@ namespace System.Speech.Internal.SrgsCompiler
                     break;
 
                 case VarEnum.VT_BOOL:
-                    _varInt = (bool)property._comValue ? unchecked((int)0xffff) : 0;
+                    _varInt = (bool)property._comValue ? unchecked(0xffff) : 0;
                     break;
 
                 case VarEnum.VT_R8:
